Comment out some codes to make Web service "works".

+++ CrowdSource/src/com/CMPUT301F12T07/crowdsource/ViewTaskActivity.java
@@ -82,7 +82,7 @@ public class ViewTaskActivity extends Activity {
         			public void onClick(DialogInterface dialog,int id) {
         				// if this button is clicked, delete the task and leave the activity
         				try {
-							db.deleteTask((int)currentTask.get_tid());
+							db.deleteTask(currentTask.get_tid());
 						} catch (Exception e) {
 							e.printStackTrace();
 						}
